refactor: choose extractor in exportProcess + extract process in Extractor.launch()
