+++ src/tab/exportProcessor/scraper/Extractor.js
@@ -1,12 +1,24 @@
+import converter from "../formatter/formatMarkdown";
+import {formatFilename} from "../formatter/formatText";
+import appInfos from "../../../infos.json";
+
 export class Extractor {
   constructor(domain) {
     this.domain = domain;
   }
 
-  extractMetadata() {
-    return {
-      title: this.getPageTitle(),
-      source: this.getPageSource()
+  async launch() {
+    try {
+      this.applyExtractorRules();
+      const markdownContent = await this.extractPage(converter[`formatMarkdown`]) //await dynamicCall(extractPages, `extract${domain.name}Page`, converter[`formatMarkdown`]);
+      const metadata = this.extractMetadata() //await dynamicCall(extractMetadata, `extract${domain.name}Metadata`);
+      const fileName = formatFilename(metadata.title, metadata.source);
+      return {markdownContent, title: metadata.title, fileName};
+    }
+    catch (e) {
+      console.error(e);
+      alert(`Error while exporting page.\n\nPlease contact me at ${appInfos.CONTACT_EMAIL} with these information if the problem persists:\n≫ The steps to reproduce the problem\n≫ The URL of this page\n≫ The app version: ${appInfos.APP_VERSION}\n≫ Screenshots illustrating the problem\n\nThank you!`);
+      return null;
     }
   }
 
@@ -16,6 +28,13 @@ export class Extractor {
    */
   async extractPage(format) {
     throw new Error("Not implemented");
+  }
+
+  extractMetadata() {
+    return {
+      title: this.getPageTitle(),
+      source: this.getPageSource()
+    }
   }
 
   /**
